Fix for Jekyll update
There seems to have been some changes in the way Jekyll assigns a
layout to a page in the presence of “defaults” in _config.yml

Before, it seems if a page layout was set, it overrode what was
specified in defaults. Now, the default overrides even an explicitly
set layout.